Fix typo in module name
