+++ app/src/main/java/org/xwiki/android/sync/activities/SyncSettingsActivity.kt
@@ -28,7 +28,7 @@ import org.xwiki.android.sync.*
 import org.xwiki.android.sync.ViewModel.SyncSettingsViewModel
 import org.xwiki.android.sync.ViewModel.SyncSettingsViewModelFactory
 import org.xwiki.android.sync.bean.ObjectSummary
-import org.xwiki.android.sync.bean.SerachResults.CustomObjectsSummariesContainer
+import org.xwiki.android.sync.bean.SearchResults.CustomObjectsSummariesContainer
 import org.xwiki.android.sync.bean.XWikiGroup
 import org.xwiki.android.sync.contactdb.UserAccount
 import org.xwiki.android.sync.contactdb.clearOldAccountContacts
