+++ app/src/main/java/org/xwiki/android/sync/rest/XWikiHttp.java
@@ -28,7 +28,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import org.xwiki.android.sync.bean.ObjectSummary;
-import org.xwiki.android.sync.bean.SerachResults.CustomObjectsSummariesContainer;
+import org.xwiki.android.sync.bean.SearchResults.CustomObjectsSummariesContainer;
 import org.xwiki.android.sync.bean.XWikiUserFull;
 
 import java.io.IOException;
